requirements.txt

# AI and Machine Learning
openai
langchain==0.3.24
langchain-community==0.3.22
langchain-openai==0.3.14
langchain-google-genai==2.0.10
google-genai==1.11.0
google-generativeai==0.8.5
# Documentation generation
python-docx==1.1.2
# Configuration and environment
python-dotenv==1.1.0
pydantic==2.11.3
pydantic-settings==2.9.1
 
# Data formatting/parsing
PyYAML
 
# HTTP clients
requests==2.32.3
httpx==0.28.1
# Utilities
tqdm==4.67.1

transformers==4.48.3
asgiref==3.5.0
backports.zoneinfo==0.2.1
cycler==0.11.0
Django==4.0.3
fonttools==4.31.2
imutils==0.5.4
kiwisolver==1.4.0
matplotlib==3.5.1
numpy
opencv-python==4.5.5.64
packaging
pandas==1.4.1
Pillow==9.0.1
psycopg2==2.9.3
pyparsing==3.0.7
pytesseract==0.3.9
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
django-tailwind
boto3
djangorestframework
